requirements.txt
aiogram>=3.0
discord.py>=2.0
fastapi>=0.100
uvicorn>=0.20
